# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tenant": {
      "get": {
        "operationId": "TenantController_current",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantContextDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found — or not yours, which is answered the same way.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current tenant profile, caller role (key and name) and permissions",
        "tags": ["tenant"]
      }
    },
    "/v1/tenant/staff": {
      "get": {
        "operationId": "TenantController_staff",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffMemberDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List staff members with roles",
        "tags": ["tenant"]
      },
      "post": {
        "operationId": "TenantController_inviteStaff",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitedStaffDto"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed: a field is missing or invalid (`message` lists them).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite a staff member (creates the tenant account if needed)",
        "tags": ["tenant"]
      }
    },
    "/v1/tenant/staff/{userId}": {
      "patch": {
        "operationId": "TenantController_updateStaff",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffChangeDto"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed: a field is missing or invalid (`message` lists them).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found — or not yours, which is answered the same way.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a staff member's role or status",
        "tags": ["tenant"]
      }
    },
    "/v1/tenant/permissions": {
      "get": {
        "operationId": "TenantController_permissions",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionInfoDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Permission catalog (platform-wide, fixed)",
        "tags": ["tenant"]
      }
    },
    "/v1/tenant/roles": {
      "get": {
        "operationId": "TenantController_roles",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleSummaryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List roles with their permissions and member counts",
        "tags": ["tenant"]
      },
      "post": {
        "operationId": "TenantController_createRole",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedRoleDto"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed: a field is missing or invalid (`message` lists them).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a custom role",
        "tags": ["tenant"]
      }
    },
    "/v1/tenant/roles/{key}": {
      "patch": {
        "operationId": "TenantController_updateRole",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedDto"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed: a field is missing or invalid (`message` lists them).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found — or not yours, which is answered the same way.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a role's name and/or permissions",
        "tags": ["tenant"]
      },
      "delete": {
        "operationId": "TenantController_deleteRole",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedDto"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed: a field is missing or invalid (`message` lists them).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found — or not yours, which is answered the same way.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an unused custom role",
        "tags": ["tenant"]
      }
    },
    "/v1/tenant/audit": {
      "get": {
        "operationId": "TenantController_audit",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "The tenant of the account's token (`tid` claim)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the credentials, code or token are wrong. Always the same answer, whatever the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recent audit records for the tenant",
        "tags": ["tenant"]
      }
    }
  },
  "info": {
    "title": "inova Core API",
...
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid credentials",
            "description": "A list for 400 validation errors, one line per field; a sentence otherwise"
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": ["statusCode", "message", "error"]
      },
...
      "TenantContextDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantSummaryDto"
          },
          "role": {
            "type": "string",
            "example": "manager",
            "description": "The caller's role key"
          },
          "roleName": {
            "type": "string",
            "example": "Домоуправител",
            "description": "The role's name as the organisation stored it"
          },
          "permissions": {
            "example": ["tenant.read", "staff.read"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["tenant", "role", "roleName", "permissions"]
      },
      "StaffMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "roleKey": {
            "type": "string",
            "example": "manager"
          },
          "status": {
            "type": "string",
            "enum": ["invited", "active", "suspended", "revoked"],
            "example": "active"
          },
          "fullName": {
            "type": "string",
            "example": "Мария Стоянова"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "maria@inova.bg"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+359881234567"
          },
          "since": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T09:30:00.000Z",
            "description": "When the membership was created"
          }
        },
        "required": ["userId", "roleKey", "status", "fullName", "email", "phone", "since"]
      },
      "InviteStaffDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "nikol@bloksofia.bg"
          },
          "fullName": {
            "type": "string",
            "example": "Nikol Petrova"
          },
          "phone": {
            "type": "string",
            "example": "+359881234567"
          },
          "roleKey": {
            "type": "string",
            "example": "manager"
          }
        },
        "required": ["email", "fullName", "roleKey"]
      },
      "InvitedStaffDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "roleKey": {
            "type": "string",
            "example": "manager"
          },
          "status": {
            "type": "string",
            "enum": ["invited", "active", "suspended", "revoked"],
            "example": "active"
          },
          "fullName": {
            "type": "string",
            "example": "Николь Петрова"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "nikol@bloksofia.bg"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "inviteSent": {
            "type": "boolean",
            "description": "False when the account was already active: it simply gains the role"
          }
        },
        "required": ["userId", "roleKey", "status", "fullName", "email", "phone", "inviteSent"]
      },
      "UpdateStaffDto": {
        "type": "object",
        "properties": {
          "roleKey": {
            "type": "string",
            "example": "manager"
          },
          "status": {
            "type": "string",
            "enum": ["active", "suspended", "revoked"]
          }
        }
      },
      "StaffChangeDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "roleKey": {
            "type": "string",
            "example": "manager"
          },
          "status": {
            "type": "string",
            "enum": ["invited", "active", "suspended", "revoked"],
            "example": "active"
          }
        },
        "required": ["userId", "roleKey", "status"]
      },
      "PermissionInfoDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "property.read"
          },
          "description": {
            "type": "string",
            "example": "View buildings and properties"
          }
        },
        "required": ["key", "description"]
      },
...
    }
  }
}